+++ src/gpgmm/d3d12/JSONSerializerD3D12.cpp
@@ -190,7 +190,7 @@ namespace gpgmm::d3d12 {
         dict.AddItem("SizeInBytes", desc.SizeInBytes);
         dict.AddItem("HeapOffset", desc.HeapOffset);
         dict.AddItem("OffsetFromResource", desc.OffsetFromResource);
-        dict.AddItem("Method", desc.Method);
+        dict.AddItem("Method", static_cast<uint32_t>(desc.Method));
         if (desc.DebugName != nullptr) {
             dict.AddItem("DebugName", WCharToUTF8(desc.DebugName));
         }
